% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{csde_logdens}
\alias{csde_logdens}
\title{Log-density of cSDE observations.}
\usage{
csde_logdens(Xt, dt, theta, mu_fun, sigma_fun, gamma_fun)
}
\arguments{
\item{Xt}{A vector of cSDE observations.}

\item{dt}{The interobservation time between each observation of the cSDE.}

\item{theta}{List of parameters of the cSDE}

\item{mu_fun}{Drift term of the cSDE which is a function of Xt and theta.}

\item{sigma_fun}{Stochastic deviationof the cSDE which is a function of theta.}

\item{gamma_fun}{Autocorrelation of the cSDE which is a function of theta, dt and N.}
}
\value{
A scalar containing the log-density of the cSDE evaluated at its arguments.
}
\description{
Calculates the log-density of `p(Xt | theta)`, where `Xt` are observations of a cSDE recorded at interobservation time `dt`.
Calculate log density of cSDE observations.
}
\details{
Calculates the log-density of `p(Xt | theta)`, where `Xt` are observations of a cSDE recorded at interobservation time `dt`. Thanks to mlysy for writing this function!
}
