% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{csde_sim}
\alias{csde_sim}
\title{Generate cSDE observations.}
\usage{
csde_sim(N, dt, X0, theta, mu_fun, sigma_fun, gamma_fun, fft = TRUE)
}
\arguments{
\item{N}{Number of observations.}

\item{dt}{The interobservation time between each observation of the cSDE.}

\item{X0}{Initial cSDE value at time `t = 0`.}

\item{theta}{List of parameters of the cSDE.}

\item{mu_fun}{Drift term of the cSDE which is a function of Xt and theta.}

\item{sigma_fun}{Stochastic deviationof the cSDE which is a function of theta.}

\item{gamma_fun}{Autocorrelation of the cSDE which is a function of theta, dt and N.}

\item{fft}{Whether to use fast (but sometimes less stable) FFT simulation method.  See [SuperGauss::rnormtz()].}
}
\value{
A vector of `N+1` cSDE observations recorded at intervals of `dt` starting from `X0`.
}
\description{
Generate cSDE observations.
}
\details{
Function provided by Prof. Lysy that simulates data generation for any coloured-noise stochastic differential equation.Tthanks to mlysy for writing this function!
}
