% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{fit_fOU_multiple_K}
\alias{fit_fOU_multiple_K}
\title{Fit multiple levels of an fOU process.}
\usage{
fit_fOU_multiple_K(Xt, Ks)
}
\arguments{
\item{Xt}{A vector of fOU observations.}

\item{Ks}{is vector of levels for the k-level approximations to be fitted.}
}
\value{
A list of `Ks` samples and stanfit objects that are fitted to `c(X0,Xt)` for each level k.
}
\description{
Fits an fOU process to the given data across multiple k-level approximations using a pre-compiled stan mode.
}
