% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{fit_fOU_process}
\alias{fit_fOU_process}
\title{Fit an fOU process.}
\usage{
fit_fOU_process(fOU_data, K, iter = 2000, control = list(adapt_delta = 0.95))
}
\arguments{
\item{fOU_data}{is a list of
\describe{
  \item{Xt}{Vector of observations at level k.}
  \item{delta_t}{The interobservation time between each observation of the fOU process.}
  \item{X0}{Initial fOU value at time `t = 0`.}
  }}

\item{K}{The level of euler-approximation which is a positive number less than N.}

\item{iter}{Maximum number of total iterations.}

\item{control}{List of control parameter for rstan::sampling. See [rstan::sampling()] for more details.}
}
\value{
A stanfit object that is fitted to `c(X0,Xt)`.
}
\description{
Fits an fOU process to the given data using a pre-compiled stan mode.
}
