% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{fou_gamma}
\alias{fou_gamma}
\title{Autocorrelation function for the fOU process.}
\usage{
fou_gamma(theta, dt, n)
}
\arguments{
\item{theta}{is a list with the parameters of the fOU process (where sigma is permanently 1):
\describe{
  \item{H}{Hurst index which is between 0 and 1.}
  \item{mu}{Mean of the fOU process which is positive.}
  \item{gamma}{Mean reverting speed of the process which is positive.}
  }}

\item{dt}{is the interobservation time between each datapoint of the fOU process.}

\item{n}{Number of observations of the fOU process.}
}
\value{
An `n` dimensional vector of covariances.
}
\description{
Autocorrelation function for the fOU process.
}
\details{
The toeplitz matrix of the autocorrelation function of the fOU process is defined by the following vector of covariances:
```
V[i] = (abs(i)^(2H) + abs(i - 2)^(2H) - 2 x abs(i - 1)^(2H)

```
}
