% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{fou_logdens}
\alias{fou_logdens}
\title{Log-density of fOU observations.}
\usage{
fou_logdens(Xt, delta_t, theta)
}
\arguments{
\item{Xt}{A vector of fOU observations.}

\item{delta_t}{is the interobservation time between each observation of the fOU process.}

\item{theta}{is a list with the parameters of the fOU process fOU process (where sigma is permanently 1):
\describe{
  \item{H}{Hurst index which is between 0 and 1.}
  \item{mu}{Mean of the fOU process which is positive.}
  \item{gamma}{Mean reverting speed of the process which is positive.}
  }}
}
\value{
A scalar containing the log-density of the fOU evaluated at its arguments.
}
\description{
Log-density of fOU observations.
}
