% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{fou_mu}
\alias{fou_mu}
\title{Drift term for the fOU process.}
\usage{
fou_mu(y_n, theta)
}
\arguments{
\item{y_n}{is a real number.}

\item{theta}{is a list with the parameters of the fOU process (where sigma is permanently 1):
\describe{
  \item{H}{Hurst index which is between 0 and 1.}
  \item{mu}{Mean of the fOU process which is greater than 0.}
  \item{gamma}{Mean reverting speed of the process which is greater than 0.}
}}
}
\value{
A real number.
}
\description{
Drift term for the fOU process.
}
