% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{fOU_predict}
\alias{fOU_predict}
\title{Make predictions using an fOU fit.}
\usage{
fOU_predict(fit, X0, delta_t, n_points, n_samples)
}
\arguments{
\item{fit}{Stan fit object of an fOU process.}

\item{X0}{Initial fOU value at time `t = 0`.}

\item{delta_t}{Interobservation time between each observation of the fOU process.}

\item{n_points}{Number of points to predict.}

\item{n_samples}{Number of sample paths to simulate.}
}
\value{
Vector of n_points containing the predicted fOU points.
}
\description{
Given a Stan fit object, it predicts the next n data points using the parameters' posterior distributions.
}
\details{
Draws parameters from the posterior distributions and simulates fOU data. The new fOU data is then aggregated to produce prediction confidence intervals. This is an expensive computation.
}
