% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{fou_sigma}
\alias{fou_sigma}
\title{Stochastic deviation for the fOU process.}
\usage{
fou_sigma(theta, dt, N)
}
\arguments{
\item{theta}{is a list with the parameters of the fOU process (where sigma is permanently 1):
\describe{
  \item{H}{Hurst index which is between 0 and 1.}
  \item{mu`}{Mean of the fOU process which is positive.}
  \item{gamma}{Mean reverting speed of the process which is positive.}
  }}

\item{dt}{is the interobservation time between each datapoint of the fOU process.}

\item{N}{Number of observations of the fOU process.}
}
\value{
A real number.
}
\description{
Stochastic deviation for the fOU process.
}
\details{
Maintained constant at 1 to simplify calculations.
}
