% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{fOU_sim}
\alias{fOU_sim}
\title{Generate fOU observations.}
\usage{
fOU_sim(N, theta, X0, delta_t)
}
\arguments{
\item{N}{Number of observations to generate.}

\item{theta}{is a list with the parameters of the fOU process fOU process (where sigma is permanently 1):
\describe{
  \item{H}{Hurst index which is between 0 and 1.}
  \item{mu}{Mean of the fOU process which is positive.}
  \item{gamma}{Mean reverting speed of the process which is positive.}
  }}

\item{X0}{Initial fOU value at time `t = 0`.}

\item{delta_t}{The interobservation time between each observation of the fOU process.}
}
\value{
A vector of `N+1` fOU observations recorded at intervals of `dt` starting from `X0`.
}
\description{
Generate fOU observations.
}
\details{
Generates data for the fOU which is characterized by the following equation:
\deqn{
dX_t = \gamma(X_t - \mu)*\Delta t + \sigma B^{H}_t
}
}
