% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-test.R
\name{logpost}
\alias{logpost}
\title{Log Posterior of test function.}
\usage{
logpost(mu, y)
}
\arguments{
\item{mu}{Mean of the observations.}

\item{y}{Vector of observations.}
}
\value{
A scalar of containing log posterior evaluated at its arguments.
}
\description{
Log posterior of given observations.
}
