% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-test.R
\name{logpost_grad}
\alias{logpost_grad}
\title{Gradient of the Log Posterior of test function.}
\usage{
logpost_grad(mu, y)
}
\arguments{
\item{mu}{Mean of the observations.}

\item{y}{Vector of observations.}
}
\value{
A scalar of containing gradient of the log posterior evaluated at its arguments.
}
\description{
Gradient of the log posterior of given observations.
}
