% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-test.R
\name{pex_acf}
\alias{pex_acf}
\title{Autocorrelation function of the Power Exponential Model}
\usage{
pex_acf(t, lambda, rho, sigma)
}
\arguments{
\item{t}{Vector of time steps of the decay being modelled.}

\item{lambda}{Scaling factor.}

\item{rho}{Autocorrelation decay parameter.}

\item{sigma}{Variance.}
}
\value{
A vector of autocorrelations based on the arguments
}
\description{
Autocorrelation function of the Power Exponential Model
}
