% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{plot_CI}
\alias{plot_CI}
\title{Plot confindence intervals.}
\usage{
plot_CI(fOU_data, fit, K, N, title = "Xt and 99\% CIs of interpolated points")
}
\arguments{
\item{fOU_data}{is a list of
\describe{
  \item{Xt}{Vector of observations at level k.}
  \item{delta_t}{The interobservation time between each observation of the fOU process.}
  \item{X0}{Initial fOU value at time `t = 0`.}
  }}

\item{fit}{Stan fit object of an fOU process.}

\item{K}{The level of euler-approximation which is a positive number less than N.}

\item{N}{Number of observations of the fOU process.}

\item{title}{Title of the plot.}
}
\value{
Plot of confidence intervals of each of the fOU parameters.
}
\description{
Given fOU_data, uses fOU_predict to generate confidence intervals, then plots n sample paths from the `true`` fOU parameters.
}
