% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{plot_likelihoods}
\alias{plot_likelihoods}
\title{Plot likelihoods of fOU parameters.}
\usage{
plot_likelihoods(
  fOU_data,
  fit,
  K,
  thresholds = list(mu = 0.1, gamma = 0.1, H = 0.05)
)
}
\arguments{
\item{fOU_data}{is a list of
\describe{
  \item{Xt}{Vector of observations at level k.}
  \item{delta_t}{The interobservation time between each observation of the fOU process.}
  \item{X0}{Initial fOU value at time `t = 0`.}
  }}

\item{fit}{Stan fit object fitted on the `fOU_data`.}

\item{K}{The level of euler-approximation which is a positive number less than N.}

\item{thresholds}{List of thresholds for mu, gamma and H to perturb each term to grid search over.}
}
\value{
Plot of the likelihoods of each parameter.
}
\description{
Plots log likelihoods for each of the fOU parameters - H, gamma and mu.
}
\details{
Check that theta_hat is a global AND local maximum.This is done by essentially grid searching one parameter while holding everything else constant.This is expensive, but feasible since our parameter constraints are fairly tight.
}
