% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{plot_param_posterior_distributions}
\alias{plot_param_posterior_distributions}
\title{Plot posterior distribution for parameters.}
\usage{
plot_param_posterior_distributions(post_tb)
}
\arguments{
\item{post_tb}{Tibble of posterior samples.}
}
\value{
Plot of the posterior distribution of each parameter of the fOU process.
}
\description{
Given a tibble of posterior samples (from fit_fOU_process), calculate the posterior distribution and plot them.
}
