% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fOU_functions.R
\name{plot_prediction_interval}
\alias{plot_prediction_interval}
\title{Plot prediction intervals.}
\usage{
plot_prediction_interval(fOU_data, fit, delta_t, n_points, n_samples)
}
\arguments{
\item{fOU_data}{is a list of
\describe{
  \item{Xt}{Vector of observations at level k.}
  \item{delta_t}{The interobservation time between each observation of the fOU process.}
  \item{X0}{Initial fOU value at time `t = 0`.}
  }}

\item{fit}{Stan fit object of an fOU process.}

\item{delta_t}{Interobservation time between each observation of the fOU process.}

\item{n_points}{Number of predicted points.}

\item{n_samples}{Number of sample paths to simulated.}
}
\value{
Plot of prediction intervals of each of the fOU parameters.
}
\description{
Given fOU_data, uses fOU_predict to generate confidence intervals, then plots n sample paths from the `true`` fOU parameters.
}
