% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-test.R
\name{test_fun}
\alias{test_fun}
\title{Test function.}
\usage{
test_fun(x)
}
\arguments{
\item{x}{A real number.}
}
\value{
A real number.
}
\description{
A test function implemented to test custom gradient functionality.
}
\details{
The default test function is `f(x) = sin(x) + x`.
}
