% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-test.R
\name{test_fun_grad}
\alias{test_fun_grad}
\title{Gradient of test function.}
\usage{
test_fun_grad(x)
}
\arguments{
\item{x}{A real number.}
}
\value{
A real number between 0 and 2.
}
\description{
Gradient of the test function implemented to test custom gradient functionality.
}
\details{
The gradient of default test function is `f'(x) = cos(x) + 1`.
}
