% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testproject1-package.R
\docType{package}
\name{testproject1-package}
\alias{testproject1-package}
\alias{testproject1}
\title{The 'testproject1' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.19.3. https://mc-stan.org
}
